requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
decorator==4.4.2
feedparser==6.0.2
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
sgmllib3k==1.0.0
six==1.15.0
tornado==6.1
tzlocal==2.1
